server/requirements-dev.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
httpx==0.25.2
faker==20.1.0
factory-boy==3.3.0
sqlalchemy-utils==0.41.1
alembic==1.13.1
pg8000==1.30.1